moove functions to utils file
