+++ butler/src/app/v2/core/integrations/spinnaker/templates/deployment/delete-unused-stage.ts
@@ -16,10 +16,11 @@
 
 import { Stage } from '../../interfaces/spinnaker-pipeline.interface'
 import { ISpinnakerConfigurationData } from '../../../../../../v1/api/configurations/interfaces'
-import { CdConfiguration, Component } from '../../../../../api/deployments/interfaces'
+import { CdConfiguration } from '../../../../../api/deployments/interfaces'
+import { DeploymentComponent } from '../../../../../api/deployments/interfaces/deployment.interface'
 
 export const getDeleteUnusedStage = (
-  component: Component,
+  component: DeploymentComponent,
   configuration: CdConfiguration,
   stageId: number,
   evalStageId: number,
