+++ butler/src/tests/v2/unit/utils/replaced-components.spec.ts
@@ -1,19 +1,10 @@
-// circulo 1:
-// Componentes: A:v1, B:v1, C:v1
-
-// Componentes: A:v1, B:v2
-// B:v1, C:v1
-
-import { flatMap, uniq, uniqBy } from 'lodash'
 import { CdConfigurationEntity } from '../../../../app/v1/api/configurations/entity'
 import { CdTypeEnum } from '../../../../app/v1/api/configurations/enums'
 import { GitProvidersEnum } from '../../../../app/v1/core/integrations/configuration/interfaces'
 import { ClusterProviderEnum } from '../../../../app/v1/core/integrations/octopipe/interfaces/octopipe-payload.interface'
 import { ComponentEntityV2 } from '../../../../app/v2/api/deployments/entity/component.entity'
 import { DeploymentEntityV2 } from '../../../../app/v2/api/deployments/entity/deployment.entity'
-import { Component } from '../../../../app/v2/api/deployments/interfaces/component.interface'
-import { Deployment, DeploymentComponent } from '../../../../app/v2/api/deployments/interfaces/deployment.interface'
-import { OctopipeDeployment } from '../../../../app/v2/core/integrations/octopipe/interfaces/octopipe-deployment.interface'
+import { componentsToBeRemoved } from '../../../../app/v2/core/integrations/utils/deployment.utils'
 
 it('must replace new components version', async() => {
   const cdConfiguration = new CdConfigurationEntity(
@@ -101,7 +92,7 @@ it('must replace new components version', async() => {
     return c
   })
 
-  const replacedVersions = replaceVersions(deployment, activeComponents)
+  const removedVersions = componentsToBeRemoved(deployment, activeComponents)
   const expectedComponents = [
     new ComponentEntityV2(
       'helm.url',
@@ -131,7 +122,7 @@ it('must replace new components version', async() => {
       null
     )
   ]
-  expect(replacedVersions.map((c) => {
+  expect(removedVersions.map((c) => {
     return {
       name: c.name,
       imageUrl: c.imageUrl
@@ -143,47 +134,3 @@ it('must replace new components version', async() => {
     }
   }))
 })
-
-
-
-
-const replaceVersions = (deployment: Deployment, activeComponents: Component[]): DeploymentComponent[] => {
-  const circleId = deployment.circleId
-  return activeComponents.filter(c => {
-    return non(deployment.components, c, circleId) || updated(deployment.components, c, circleId)
-    return !deployment.components?.some(dc => nonExistendConditions(dc, c, circleId)) || deployment.components?.some(dc => oldVersionCondition(dc, c, circleId))
-  })
-  const removedComponents = activeComponents.filter(c => !deployment.components?.some(dc => nonExistendConditions(dc, c, circleId)))
-  const updatedComponents = activeComponents.filter(c => deployment.components?.some(dc => oldVersionCondition(dc, c, circleId)))
-  return updatedComponents.concat(removedComponents)
-}
-
-const non = (deploymentComponents: DeploymentComponent[] | undefined, activeComponent: Component, circleId: string) => {
-  return !deploymentComponents?.some(dc => nonExistendConditions(dc, activeComponent, circleId))
-}
-
-const updated = (deploymentComponents: DeploymentComponent[] | undefined, activeComponent: Component, circleId: string) => {
-  return deploymentComponents?.some(dc => oldVersionCondition(dc, activeComponent, circleId))
-}
-
-const nonExistendConditions = (deploymentComponent: DeploymentComponent, activeComponent: Component, circleId: string): boolean => {
-  return isSameName(deploymentComponent, activeComponent) && isSameCircle(circleId, activeComponent.deployment)
-}
-
-const oldVersionCondition = (deploymentComponent: DeploymentComponent, activeComponent: Component, circleId: string): boolean => {
-  return isSameNameAndDifferenteVersion(deploymentComponent, activeComponent) && isSameCircle(circleId, activeComponent.deployment)
-}
-
-const isSameNameAndDifferenteVersion = (deploymentComponent: DeploymentComponent, activeComponent: Component): boolean => {
-  return isSameName(deploymentComponent, activeComponent) && deploymentComponent.imageUrl !== activeComponent.imageUrl
-}
-
-const isSameName = (deploymentComponent: DeploymentComponent, activeComponent: Component): boolean => {
-  return deploymentComponent.name === activeComponent.name
-}
-
-const isSameCircle = (circleId: string, deployment: Deployment): boolean => {
-  return circleId === deployment.circleId
-}
-
-
